Final boss jaa

+++ Main/deposit.cpp
@@ -5,7 +5,7 @@
 #include "include/runCreate_img.h"
 #include <string>
 
-int depositMoney(std::string username, std::string password, double amount)
+int depositMoney(std::string username, std::string password,long double amount)
 {
     loadUsers();
