+++ Main/withdraw.cpp
@@ -9,7 +9,7 @@
 
 using namespace std;
 
-int withdrawMoney(std::string username,std::string password, double amount, double &newBalance)
+int withdrawMoney(std::string username,std::string password, long double amount, double &newBalance)
 {
     loadUsers();
